envoy: extract HttpConnectionManager
